+++ src/stats_duck_extension.cpp
@@ -8,6 +8,7 @@
 #include "summary_stats_function.hpp"
 #include "correlation_function.hpp"
 #include "rank_correlation_function.hpp"
+#include "sign_test_function.hpp"
 #include "normality_function.hpp"
 #include "anova_function.hpp"
 #include "chisq_function.hpp"
@@ -30,6 +31,7 @@ static void LoadInternal(ExtensionLoader &loader) {
 	RegisterPearsonTest(loader);
 	RegisterSpearmanTest(loader);
 	RegisterKendallTest(loader);
+	RegisterSignTest(loader);
 	RegisterJarqueBera(loader);
 	RegisterAnovaOneway(loader);
 	RegisterChiSquareTests(loader);